feat: add exception on user feature

+++ src/main/java/com/tickets/ticketmanagement/users/service/impl/UserServiceImpl.java
@@ -1,11 +1,14 @@
 package com.tickets.ticketmanagement.users.service.impl;
 
 import java.util.List;
-import java.util.Optional;
 
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
+import com.tickets.ticketmanagement.exception.DataNotFoundException;
+import com.tickets.ticketmanagement.exception.DatabaseOperationException;
+import com.tickets.ticketmanagement.exception.UserAlreadyExistException;
 import com.tickets.ticketmanagement.roles.entity.RolesEntity;
 import com.tickets.ticketmanagement.roles.repository.RolesRepository;
 import com.tickets.ticketmanagement.users.dto.RegisterRequestDto;
@@ -29,6 +32,9 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public User register(RegisterRequestDto registerRequestDto) {
+        if (userRepository.findByEmail(registerRequestDto.getEmail()).isPresent()) {
+            throw new UserAlreadyExistException("user with email " + registerRequestDto.getEmail() + " already exists");
+        }
         User user = new User();
         user.setName(registerRequestDto.getName());
         user.setEmail(registerRequestDto.getEmail());
@@ -36,17 +42,22 @@ public class UserServiceImpl implements UserService {
         RolesEntity roles = new RolesEntity();
         roles.setId(registerRequestDto.getRole());
         user.setRole(roles);
-        return userRepository.save(user);
+
+        try {
+            return userRepository.save(user);
+        } catch (DataIntegrityViolationException ex) {
+            throw new DatabaseOperationException("Failed to register user due to database error.", ex);
+        }
     }
 
     @Override
     public User findByEmail(String email) {
-        return userRepository.findByEmail(email).orElse(null);
+        return userRepository.findByEmail(email).orElseThrow(() -> new DataNotFoundException("user with email " + email + " not found"));
     }
 
     @Override
     public User findById(Long id) {
-        return userRepository.findById(id).orElseThrow(()-> new RuntimeException("user with id " + id + "not found"));
+        return userRepository.findById(id).orElseThrow(()-> new DataNotFoundException("user with id " + id + "not found"));
     }
 
     @Override
@@ -56,15 +67,20 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public Void deleteBy(Long id) {
-        userRepository.deleteById(id);
+        if (userRepository.existsById(id)) {
+            throw new DataNotFoundException("user with id " + id + " not found");
+        }
+        try {
+            userRepository.deleteById(id);
+        } catch (DataIntegrityViolationException ex ) {
+            throw new DatabaseOperationException("Failed to delete user due to database error", ex);
+        }
         return null; 
     }
 
     @Override
     public User updateProfile(Long id, UserProfileUpdateDto updateProfileDto) {
-        Optional<User> optionalUser = userRepository.findById(id);
-        if (optionalUser.isPresent()) {
-            User user = optionalUser.get();
+        User user = userRepository.findById(id).orElseThrow(() -> new DataNotFoundException("User not found with id " + id));
             user.setName(updateProfileDto.getName());
             user.setEmail(updateProfileDto.getEmail());
             user.setPassword(updateProfileDto.getPassword());
@@ -72,18 +88,19 @@ public class UserServiceImpl implements UserService {
             Long roleId = updateProfileDto.getRoleId();
             if (roleId != null) {
                 RolesEntity role = rolesRepository.findById(roleId)
-                .orElseThrow(() -> new RuntimeException("role not found with id " + roleId));
+                .orElseThrow(() -> new DataNotFoundException("role not found with id " + roleId));
                 user.setRole(role);
             }
-            return userRepository.save(user);
-        } else {
-            throw new RuntimeException("User not found with id " + id);
-        }
+            try {
+                return userRepository.save(user);
+            } catch (DataIntegrityViolationException ex) {
+                throw new DatabaseOperationException("Failed to update user profile due to database error.", ex);
+            }
     }
 
     @Override
     public User findByRefferalCode(String refferalCode) {
-        return userRepository.findByEmail(refferalCode).orElse(null);
+        return userRepository.findByEmail(refferalCode).orElseThrow(()-> new DataNotFoundException("user with referral id " + refferalCode + "not found"));
     }
 
 }
